Base functions working, need more UI and pubsub
Dont even know how the fk the modules would talk anymore since they
are separate. Have a decent idea from the library project, but feels
like it gets complicated. I think it would just be way easier
with a pubsub, so I'll do that. Base pubsub seems simple enough
as well. Need to also make the rest of the UI and hook them up to
the functions.