Add simple sorting method on all header cells
can maybe take out 'description header'. Will try to give it a
short display limit of 3 lines or so of width X. Need to re-look
at how to make a text string use the elipsis technique(...) to
display it with those 3 dots at the end.

Think about where to put the 'edit' button. Will prbably make any
cell clickable, so that any of the cells for a given todo-item
will give you the 'edit todo' prompt.

Also add Project Name edit prompt, AND display the Project name on top
of the table probably (maybe put the edit name option right next to
this title that goes above the table)