Add 'Todo table' with ALL sorting. Need edit modal
Table highlights each table row when its hovered. Clicking on the
table headers will sort them by that field. Date column recently
changed from sorting by Date created to sorting by Due Date. Clicking
on a row will display the name of the item in that row

Still need to add an Edit modal. 99% sure that it should be separate
from the 'create todo' modal. This is because if someone exits editing
'modal A' without saving, and then they click 'create todo'. They should
be able to create a new Todo from blank, and at the same time, once
they submit OR exit the create Todo, they shouls still be able to
go back to editing 'modal A'. Only when you select another modal to
EDIT, should the 'edit modal' be set to blank.

To achieve this i believe we need an outside 'selected Todo' variable.
This way, we can check if the next todo item to prompt the 'edit Modal'
is different than the last one. Only if its different, set the
Edit Modal to blank.
Also use html templates, this is probably better. Either use a separate
template than the 'create Todo', or most likely, use a single template,
make 2 copies (1 for Create, 1 for Edit), and tweak the copy of the
Edit Modal a little. Keep references cached at the top of the js
display controller. Thumbs Up.