Add simple EditTodo modal. Need to decide behavior
Created an html template for the New Todo modal. Both the Edit Todo
modal, and the New Todo modal are copy nodes of this template.

Need to decide functionality of when I want each edit Todo modal to
clear. Also on whether or not I want another modal that prompts
'Discard changes? Yes or Save', whenever you try to close the modal.

Originally I wanted the behavior in the previous commit- Edit Modal
only clears when I click on a different Todo to edit than the previous
one clicked. Else, if you close a modal then click to edit the same
one, the changed would be saved
Now I think I might want the behavior to be: Adding the 'Discard
changes prompt'. However this would look strange because the EditTodo
is already a modal, so we would probably want to change how the
EditTodo is displayed. Change the content page to display TodoInfo
clicked, like Youtube Video Manager.

Leave all this for later tho. Next up, we see if we can clean code,
and add time.js and JSON and persisting data in LOCAL storage!!